Add more actions to users resource
