+++ app/controllers/api/v1/users_controller.rb
@@ -2,18 +2,24 @@
 
 module Api
   module V1
-    # Register users
-    class UsersController < ApplicationController
+    # Manages users
+    class UsersController < AuthenticatedController
+      def show
+        render json: current_user
+      end
+
       def create
         user = User.new(user_params)
 
         if user.save
           # TODO: replace with real location
-          render json: user, status: :created, location: nil
+          render json: user, status: :created, location: api_v1_users_path
         else
           render json: user, status: :unprocessable_entity
         end
       end
+
+      def update; end
 
       private
