+++ app/models/authentication/user_by_id.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Authentication
+  # Finds a user by its ID
+  class UserById < ApplicationModel
+    def initialize(id)
+      super()
+      @id = id
+    end
+
+    def to_model
+      User.find(@id.to_s)
+    end
+  end
+end
