+++ config/routes.rb
@@ -4,7 +4,7 @@ Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
       resources :sessions, only: %i[create]
-      resources :users, only: %i[create]
+      resource :users, only: %i[show create update]
     end
   end
 end
